requirements.txt
flask
pyyaml
pandas
flask-mysql
sklearn
numpy
